fix: quantity followers and followings are double

+++ src/model/util/factory/AgentConfigFactory.java
@@ -13,13 +13,13 @@ public class AgentConfigFactory {
     }
 
 
-    public AgentConfig createAgentConfig(Agent agentData, int agentCant, int followers, int followings){
+    public AgentConfig createAgentConfig(Agent agentData, int agentCant, double followers, double followings){
         AgentConfig agentConfig = new AgentConfig(agentData, agentCant, followers, followings);
         agentsConfigs.add(agentConfig);
         return agentConfig;
     }
 
-    public AgentConfig createAgentConfig(Agent agentData, int agentCant, int followersCant, int followings, String name){
+    public AgentConfig createAgentConfig(Agent agentData, int agentCant, double followersCant, double followings, String name){
         AgentConfig agentConfig = new AgentConfig(agentData, agentCant, followersCant, followings, name);
         agentsConfigs.add(agentConfig);
         return agentConfig;
